+++ spec/dummy/app/channels/widget_channel.rb
@@ -1,0 +1,9 @@
+class WidgetChannel < ActionCable::Channel::Base
+  def subscribed
+    stream_from "widgets"
+  end
+
+  def follow(data)
+    transmit({ status: "following", id: data["id"] })
+  end
+end