+++ spec/dummy/app/jobs/aborted_job.rb
@@ -1,0 +1,6 @@
+class AbortedJob < ActiveJob::Base
+  before_perform { throw :abort }
+
+  def perform
+  end
+end